requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.10
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-editorjs==0.2.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.5.5.62
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
zope.interface==5.4.0
